Let's start the C++ half of this codebase.
Included: a working CMake file (it was hell to get this working until I
realized that the first argument to find_package() was case-sensitive!), a
"hello world"-type main.cpp that actually links to and uses OpenCV to
demonstrate that the CMake process worked, and some hastily scrawled README
updates.

The next step will be getting a config file reader.  Gotta keep it simple.