Got Boost.ProgramOptions to compile and read from the config file.
There's still a minor annoyance: I expected Boost.ProgramOptions to parse the
comma-separated list itself and fully populate the vector<string>.  It's not a
big deal; we can just use vm["network.roborio_addresses"].as<string>() for now
and be done with it.  Of course that means we have to parse on our own, which
is a shame because ProgramOptions already has a parser.

A bit of refactoring and then onto the joys of threaded networking.