Trying what's supposed to be a simpler means of opening client sockets with getaddrinfo().
The technique comes from
https://beej.us/guide/bgnet/output/html/singlepage/bgnet.html#connect.  It's
not working right now -- the connect() call returns -1 with an errno of 111,
and there's no error message that strerror_r() can associate with that with a
message string.

It's not like it matters.  I already got this to work, so we'll just use that
for now.  Besides, we still need select() so that we can code in connection
timeouts.

This branch will still be here when I need it.