Changed CMakeLists.txt and added PapasVision.cpp
I changed CMakeLists.txt because the issue is that on a windows machine
the cygwin cmake setup thinks it wants the cmake gui file so I had to
change the location. The PapasVision.cpp is being worked on. It is being
changed from Java to C++ currently to make sure that it works with the
NVIDIA Jetson TK1.