Slogging through the Java -> C++ conversion process.
We now generate the cancelcolors, cancelcolors_morphfilt, and framecontours
images.

The hardest part of the conversion?  PapasVision::findContours(), by far.
Imagine _my_ surprise to learn that the OpenCV C++ API had no MatOfPoint
class!  It took a while to find the arguments that findContours() was really
supposed to take.  All of this "InputOutputArray" nonsense is confusing, too
-- they're really sd::vector<T>s at heart, so why not just declare vectors to
be the arguments these functions actually take?