TryCatch in Severrunnable and trying to get Gyro
I addeed a try catch to get rid of some errors with the code not being
able to get the code on to the robot. The gyro has some errors with
wanting to the workd with the code. Tryging to see what the issue is
right now besides the runnable.java class.