Unbreak the Cygwin build.
Cygwin's <string.h> header _has_ strerror_r() (as does Linux, where I do my
builds), but #define-ing _GNU_SOURCE as the header recommends doesn't seem to
be enough to convince Cygwin to enable it.  Instead, we get the int-returning
XSI version.

This is too unreliable to be allowed to persist.  I'm pretty certain Cygwin's
in the wrong here, but we have to move on.  So I've turned on the
non-reentrant strerror(); fortunately, we only have one thread right now, but
if we end up with more than one, we're going to have to add a scoped_lock.