Finished XMLParser.java, added PapasData and TerminalTest.java
XMLParser.java is SAX Parser that reads XML strings and converts them into strings that Java can understand for the rest of the code can understand. PapasData.java has the result and data XML that came from PapasVision. TerminalTest.java which is a jar when building is an ant file that we can run in any form of terminal or command prompt.