Added shootWheel & feedWheel start-ups, added compatability for Sparks and CANTalons Added velocity for shootWheel & feedWheel so they can start and stop.
There is a try and catch exception block for added parts and a thread.sleep to allow for the shootWheel to accelerate before the feedWheel starts up. Added variables of time for the shootWheel to be ready to launch & for the feedWheel to be needed active. Imported Sparks and Speedcontroller for CANTalons/Sparks compatability. Changed CANTalon classes to Speedcontroller.