Trying to debug my buggy linear regression function.
I've added a "debug variables" hash map that can be used to save intermediates
even as the result of the output is directed to a here-doc.  The graph now
prints the equation in the title, and the command-line program can accept
arbitrary coordinates in the program args.

The problem's still there, though!