The curve-fitting function is fixed.
That is to say, the line of best fit goes through the center of mass.

I had to rely on a different algorithm that I still presume is Simple Linear
Regression.

Also fixed the random point generator in TerminaTest.java so it generates
something that actually resenmbles a quadratic equation.