Committing some unbuilt and non-working code that will serve as a base to build upon.
The code spawns the connecting code in a separate thread, then waits on a
condition variable for TIMEOUT milliseconds.  Either the connecting thread
gets a response and notifies the condition variable within time or it doesn't;
either way, we'll return _something_ from the main thread after the timeout
has passed.

This also gives us an opportunity to make all of the connections in the list
in parallel rather than just doing them sequentially.

Needs more work.