Added the timeouts to the config file and used them.
- config/camera-client.ini: Clean up the file a little, and add the two new
  keys for the two separate timeouts.

- include/Config.h: Add accessors for the two new timeouts.

- src/Config.cpp: Wire in and implement the two new timeouts.

- src/RemoteTransmitter.cpp: Use the config timeout functions instead of magic
  constants.

I'm still not satisfied: If we can't connect to the robot, the DriverStation
never gets notified because we deliberately quit early.  Yes, I _could_ just
reverse the order and have the driver station connect first with a briefer
timeout, but then the thought struck me: shouldn't we be continuously trying
to connect (or reconnect) to the DriverStation and the Robot _anyway_?  After
all, network brownouts are known to happen.

Instead of just connecting once at the beginning and hoping for the best, we
should have the smarts to enter the loop with trying and retrying in mind.

That's new code, though.  I'll have to think about it.