Unbreak the build, at least for Mageialinux.
The changes committed last night to get that quirky Ubuntu laptop to build
were, as it turned out, breaking changes; calling find_package() on Thread
blows up for me, and probably takes out the Cygwin build, too.

I'm not sure why, but I have a different approach for fixing it: just edit the
linker flags to force a link to pthreads for platforms -- like quirky Ubuntu
laptops -- that don't seem to do this explicitly.

We'll test it in a bit.