We're increasingly starting to realize that the peg and boiler solutions need different height measures.
So we've added a height argument to findDistToGoal(), which then required documenting the constants that we
could pass into it.  We also cleaned up main, but only a little: we still need main to do nothing but finding
peg solutions until we've got this right.  We're getting false negatives on image number 9 right now, so that's
a place to start.