Make findPeg() and findBoiler() work with string arguments, too.
The way it works right now is that you can call findBoiler(s) to find a
boiler solution on the string s, or findBoiler("") to find a boiler solution
using a camera image.

If s is a relative path, it will be relative to config.cameraFolder() (or
relative to the current directory if config.cameraFolder() is an empty
string.)  Otherwise, s is an absolute path and will be opened directly.

The writeIntermediateFilesToDisk flag is still respected, and will cause the
intermediate debugging images to be written to the config.cameraFolder() (or
the current directory, if there's no cameraFolder) when the flag is set.