With David's help, implemented a Peg rejection heuristic that doesn't produce false negatives.
The idea here is simple: divide the area of the contour by the area of
the oriented bounding rectangle (not the convex hull -- that's still
too large when dealing with irregular contours.)  If the ratio is
greater than 0.8, accept, otherwise reject.

This simple test throws away the overwhelming majority of the contours
that we need to consider -- it's actually working excellently.

Will test with more samples next.