Made the Peg solution robust against the obscuring of the tape by the *actual* peg.
More importantly, altered the distance calculation for the Peg
solution based on simple trigonometric proportions and known
constants.  It was easier to calculate this than the same calculation
for the Boiler.

We're still not using boundingBoxesAreNotBothVertical() as a
rejection heuristic.  We don't seem to need it.  We *do* need two
separate scoring criterion in order to prevent the two segements of a
peg-obscured piece of relective tape from causing duplicate scores.