Divorced the message queue from the RemoteTransmitter.
I explained why this was necessary in the TransmissionBuffer class
comment. The change permits be to turn most of those static functions from the
RemoteTransmitter into instance functions, but it did result in many more
arguments to the ThreadFunction.

With that out of the way, added the ability to send driver station strings in
the main menu.  Doing so for the robot won't be difficult, but that's another
commit.

The real challenge is going to be inducing disconnections and reconnections.