Use self-aware Connection objects as wrappers around the hodge-podge that came before.
These Connection classes can be accessed asynchronously and independently of
the RemoteTransmitter's loop, allowing the interactive menu to query the
status of the Connections and even disconnect and reconnect while the
transmission loop is taking place.

The immediate next challenge is getting shutdown to be clean, the way it used
to be.