Automatic reconnection upon disconnection works, but it's not going to mean much without handling SIGPIPE.
I also fixed the hang when the program was closing -- turns out we were
missing a ref() around the const bool& shutdown, so we weren't actually
passing in the same variable that main.cpp was modifying.

I'm close.  Dealing with write errors and SIGPIPE disconnections should make
this as robust as it has to be to survive brownouts.