Handle SIGPIPE properly.  At last!
I had originally written a custom signal handler for this, but seeing other
examples of SIGPIPE handling online led me to conclude that ignoring the
signal altogether was more elegant -- that forces write() calls to return
EPIPE, which is handled in-band and much more easily than a global signal
handler.

At this point, I cam confident that the driver station, when disconnected,
will quietly attempt to reconnect until it succeeds.  On to the RoboRIO now.