Start work on a fake XML composing menu option.
It deals with incomplete information, but it can't parse doubles right now.

We also have a problem that probably can't be solved for the moment: a small
number of messages to the robot can get dropped if the robot catastrophically
disconnects, because it takes more than one write() call to finally get that
EPIPE indicating that something went awry.  Not much that can be done about
that.