Added a simple crontab that can be built on the Jetson.
We use cmake's configure_file() feature here to ensure that the generated
crontab always contains the absolute path to the git source code tree.  The
cron job will make this path the currently directory and then run
bin/camera-client the same way we humans do.

Since crontab is a generated file, it needs to be ignored.