Take note of a couple more corpus images to address, and add a new test.
The new test is for the Peg solution, and verifies that the two bands are
"more horizontal than vertical."  We could also test for aspect ratio, but I
don't see the need to do that at this time.

It would be nice if we used the true width between the two peg bands instead
of the distance between the left bottommost point and the right bottommost
point.  What was that formula for projecting a point onto a line segment
again?