¡PELIGRO!  This code is for the test bot _only_.
Don't deploy it to the bagged bot, and don't merge this entire branch into
master!

This is a fork of the 1.0.0 tag (one of the last tags we were able to deploy
without COMM problems), with changes to Robot.java specifically intended to
make the test robot drive properly, including replacing relevant motor
controllers.  This has been confirmed to work with the switch (!), so there
may be networking changes in this branch which we will cherry-pick into
master.

Time will tell.