Separated the AutoAimCommand into an independently testable "impl" class and the actual command.
Much plumbing needed to be done, like cherry-picking a bugfix from master, or
making the JAR task build (not strictly necessary, but I thought it was at the
time and I like the side effect now.)

The TerminalTest.jar program's "listen" subcommand now fires up this "impl"
class and reports whether it rotates are not.

For now, veering to the left or right or stopping is all the impl class can do
when not given real camera data.

TODO: we should reject as erroneous vision solution types which differ from
the type we had when we were first initialized.